+++ tests/ErrorOr.SwitchTests.cs
@@ -8,50 +8,50 @@ public class SwitchTests
     private record Person(string Name);
 
     [Fact]
-    public void CallingSwitch_WhenIsSuccess_ShouldExecuteOnValueAction()
+    public void CallingSwitch_WhenIsSuccess_ShouldExecuteThenAction()
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new Person("Amichai");
-        void OnValueAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
-        void OnErrorsAction(IReadOnlyList<Error> _) => throw new Exception("Should not be called");
+        void ThenAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
+        void ElsesAction(IReadOnlyList<Error> _) => throw new Exception("Should not be called");
 
         // Act
         Action action = () => errorOrPerson.Switch(
-            OnValueAction,
-            OnErrorsAction);
+            ThenAction,
+            ElsesAction);
 
         // Assert
         action.Should().NotThrow();
     }
 
     [Fact]
-    public void CallingSwitch_WhenIsError_ShouldExecuteOnErrorAction()
+    public void CallingSwitch_WhenIsError_ShouldExecuteElseAction()
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new List<Error> { Error.Validation(), Error.Conflict() };
-        void OnValueAction(Person _) => throw new Exception("Should not be called");
-        void OnErrorsAction(IReadOnlyList<Error> errors) => errors.Should().BeEquivalentTo(errorOrPerson.Errors);
+        void ThenAction(Person _) => throw new Exception("Should not be called");
+        void ElsesAction(IReadOnlyList<Error> errors) => errors.Should().BeEquivalentTo(errorOrPerson.Errors);
 
         // Act
         Action action = () => errorOrPerson.Switch(
-            OnValueAction,
-            OnErrorsAction);
+            ThenAction,
+            ElsesAction);
 
         // Assert
         action.Should().NotThrow();
     }
 
     [Fact]
-    public void CallingSwitchFirst_WhenIsSuccess_ShouldExecuteOnValueAction()
+    public void CallingSwitchFirst_WhenIsSuccess_ShouldExecuteThenAction()
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new Person("Amichai");
-        void OnValueAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
+        void ThenAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
         void OnFirstErrorAction(Error _) => throw new Exception("Should not be called");
 
         // Act
         Action action = () => errorOrPerson.SwitchFirst(
-            OnValueAction,
+            ThenAction,
             OnFirstErrorAction);
 
         // Assert
@@ -63,14 +63,14 @@ public class SwitchTests
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new List<Error> { Error.Validation(), Error.Conflict() };
-        void OnValueAction(Person _) => throw new Exception("Should not be called");
+        void ThenAction(Person _) => throw new Exception("Should not be called");
         void OnFirstErrorAction(Error errors)
             => errors.Should().BeEquivalentTo(errorOrPerson.Errors[0])
                 .And.BeEquivalentTo(errorOrPerson.FirstError);
 
         // Act
         Action action = () => errorOrPerson.SwitchFirst(
-            OnValueAction,
+            ThenAction,
             OnFirstErrorAction);
 
         // Assert
@@ -78,34 +78,34 @@ public class SwitchTests
     }
 
     [Fact]
-    public async Task CallingSwitchFirstAfterThenAsync_WhenIsSuccess_ShouldExecuteOnValueAction()
+    public async Task CallingSwitchFirstAfterThenAsync_WhenIsSuccess_ShouldExecuteThenAction()
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new Person("Amichai");
-        void OnValueAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
+        void ThenAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
         void OnFirstErrorAction(Error _) => throw new Exception("Should not be called");
 
         // Act
         Func<Task> action = () => errorOrPerson
             .ThenAsync(person => Task.FromResult(person))
-            .SwitchFirst(OnValueAction, OnFirstErrorAction);
+            .SwitchFirst(ThenAction, OnFirstErrorAction);
 
         // Assert
         await action.Should().NotThrowAsync();
     }
 
     [Fact]
-    public async Task CallingSwitchAfterThenAsync_WhenIsSuccess_ShouldExecuteOnValueAction()
+    public async Task CallingSwitchAfterThenAsync_WhenIsSuccess_ShouldExecuteThenAction()
     {
         // Arrange
         ErrorOr<Person> errorOrPerson = new Person("Amichai");
-        void OnValueAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
-        void OnErrorsAction(IReadOnlyList<Error> _) => throw new Exception("Should not be called");
+        void ThenAction(Person person) => person.Should().BeEquivalentTo(errorOrPerson.Value);
+        void ElsesAction(IReadOnlyList<Error> _) => throw new Exception("Should not be called");
 
         // Act
         Func<Task> action = () => errorOrPerson
             .ThenAsync(person => Task.FromResult(person))
-            .Switch(OnValueAction, OnErrorsAction);
+            .Switch(ThenAction, ElsesAction);
 
         // Assert
         await action.Should().NotThrowAsync();